pyproject.toml
[project]
name = "beardantic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.26.0",
    "pyarrow>=19.0.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]
